Linear-and-logarithmic combined scale for y-scales.
Requires gnuplot 5.2.  (nonlinear)

Many plots are most interesting for small values around 0, but the scale
may be given by some few outliers.  Some remote source may be good and
thus have values close to zero, while other have much larger values.  Then
the larger determine the plot scales, making the details of the good
remote sources invisible.

Due to the values taking both positive and negative values, and also at
or close to zero, the use of a pure logarithmic scale is not possible.

Instead, an inner linear range is chosen, and then both above and below
that logarithmic scales are used.

The range of the linear scale is chosen thus: for each plotted curve
determine the 50th percentile absolute value.  This ignores
the worst outliers of the curve.  The curve of a plot with the smallest
such percentile is chosen.  The linear range is then set as the smallest
value of 10^n * (1, 2 or 5) that is just larger than the limit.

The y tic marks are forced such that 0, 0.5*limit and limit are shown,
as well as limit*10^(1, 2, 3, ... 6).  for both positive and negative
values.

No ranges are set, the gnuplot automatic range selection works, and e.g.
if only a large logarithmic part is needed on the positive or negative
side, only that is shown.